# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Effecti ERP API",
        "version": "1.0"
    },
    "paths": {
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Lista paginada",
                "operationId": "de55d50d1b708d9e5c63ee3f8fe8f54a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Ativo",
                                "Inativo"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cliente"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cria cliente - valida CPF/CNPJ",
                "operationId": "4ac6cdfa8d6c188fa2cb4f3263d4318a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteStore"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "CPF duplicado ou inválido"
                    }
                }
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Detalha",
                "operationId": "0d3826d79390929601421b5810b61231",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualiza",
                "operationId": "74956b09f3ea120b5dd531c73776ca11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Exclui - bloqueia se tiver contrato",
                "operationId": "d15e304c609752c9f409ab81072fb52e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sem conteúdo"
                    },
                    "422": {
                        "description": "Possui contratos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contratos": {
            "get": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Listar contratos paginados",
                "operationId": "11f09a10344e23955021d6249957653e",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Ativo",
                                "Cancelado"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Criar novo contrato",
                "operationId": "8baa1a8b67f0d0ff7c3e54bc36bb7f1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "data_inicio"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "data_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    },
                                    "data_termino": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Ativo",
                                            "Cancelado"
                                        ],
                                        "example": "Ativo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    }
                }
            }
        },
        "/api/contratos/{id}": {
            "get": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Visualizar contrato com itens e total calculado",
                "operationId": "7b60228bf02b931b36084ba65eb05a2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contrato encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contrato"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contrato não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Atualizar contrato",
                "operationId": "287076dcac109176c976e161dddd94d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contrato"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Excluir contrato",
                "operationId": "6dfe1836df08e16da606b639554a30a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Excluído"
                    }
                }
            }
        },
        "/api/contratos/{id}/itens": {
            "post": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Adicionar item ao contrato",
                "operationId": "de1187746adc0094d88c0959533e1ace",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "servico_id",
                                    "quantidade",
                                    "valor_unitario"
                                ],
                                "properties": {
                                    "servico_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "quantidade": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "valor_unitario": {
                                        "type": "number",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item adicionado"
                    }
                }
            }
        },
        "/api/servicos": {
            "get": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Lista paginada de serviços",
                "description": "Retorna serviços com filtro opcional por nome",
                "operationId": "16fd13f4f66c0daa1379d1949fe23b05",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filtra por nome",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Internet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Servico"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Cria um novo serviço",
                "operationId": "c28337c7fc0e5c39322dc143399e2240",
                "requestBody": {
                    "description": "Dados do serviço",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServicoStore"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Servico"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou (nome duplicado ou valor inválido)"
                    }
                }
            }
        },
        "/api/servicos/{id}": {
            "get": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Detalha um serviço",
                "operationId": "16aed17b468eae5ca1f359be70629727",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do serviço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Servico"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Atualiza serviço",
                "operationId": "0e0951239106420aeea678fbe3cc21e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServicoUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Servico"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Serviços"
                ],
                "summary": "Exclui serviço - bloqueia se estiver em uso",
                "operationId": "d6ae0ce3ca2c44964a431754aedb8dae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Excluído com sucesso"
                    },
                    "422": {
                        "description": "Serviço em uso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Serviço está vinculado a contratos e não pode ser excluído."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Cliente": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nome": {
                        "type": "string",
                        "example": "ACME LTDA"
                    },
                    "cpf_cnpj": {
                        "type": "string",
                        "example": "11222333000181"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Ativo",
                            "Inativo"
                        ]
                    },
                    "contratos_count": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ClienteStore": {
                "required": [
                    "nome",
                    "cpf_cnpj",
                    "email"
                ],
                "properties": {
                    "nome": {
                        "type": "string"
                    },
                    "cpf_cnpj": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ClienteUpdate": {
                "properties": {
                    "nome": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Ativo",
                            "Inativo"
                        ]
                    }
                },
                "type": "object"
            },
            "Contrato": {
                "title": "Contrato",
                "required": [
                    "cliente_id",
                    "data_inicio",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cliente_id": {
                        "description": "ID do cliente",
                        "type": "integer",
                        "example": 1
                    },
                    "data_inicio": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "data_termino": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Ativo",
                            "Cancelado"
                        ],
                        "example": "Ativo"
                    },
                    "valor_total": {
                        "description": "Soma dos itens (accessor)",
                        "type": "number",
                        "format": "float",
                        "example": 1500
                    },
                    "itens": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ContratoItem"
                        }
                    }
                },
                "type": "object"
            },
            "ContratoItem": {
                "title": "Item do Contrato",
                "required": [
                    "servico_id",
                    "quantidade",
                    "valor_unitario"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "contrato_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "servico_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "quantidade": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 10
                    },
                    "valor_unitario": {
                        "type": "number",
                        "format": "float",
                        "example": 150
                    }
                },
                "type": "object"
            },
            "Servico": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Internet Fibra"
                    },
                    "valor_base_mensal": {
                        "type": "number",
                        "format": "float",
                        "example": 99.9
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ServicoStore": {
                "required": [
                    "nome",
                    "valor_base_mensal"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Cloud Backup"
                    },
                    "valor_base_mensal": {
                        "type": "number",
                        "format": "float",
                        "example": 49.9
                    }
                },
                "type": "object"
            },
            "ServicoUpdate": {
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Cloud Backup Pro"
                    },
                    "valor_base_mensal": {
                        "type": "number",
                        "format": "float",
                        "example": 59.9
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Clientes",
            "description": "CRUD de clientes - Regras 1,2,3 do PDF"
        },
        {
            "name": "Contratos",
            "description": "Operações de contratos e itens"
        },
        {
            "name": "Serviços",
            "description": "CRUD de serviços - Regra: não excluir se estiver em contrato"
        }
    ]
}